pyproject.toml
[project]
name = "fastapi-sendparcel"
version = "0.1.0"
description = "FastAPI adapter for python-sendparcel"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "Dominik Kozaczko" }]
dependencies = [
    "fastapi>=0.115.0",
    "pydantic-settings>=2.0.0",
    "python-sendparcel>=0.1.0",
    "anyio>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.9.0",
]
sqlalchemy = [
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fastapi_sendparcel"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 80
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "A", "SIM", "RUF"]
ignore = ["N818", "B008", "B027"]

[tool.uv.sources]
python-sendparcel = { path = "../python-sendparcel", editable = true }
